feat(backend): add getAndUpdateEntity to the DBAdapter interface
We need to be able to get and update an entity within the same
transaction such that conflicts do not occur. This commit adds the
getAndUpdateEntity function to the DBAdapter interface, and implements
it for the PostgresAdapter. We employ row-level locking using
SELECT ... FOR UPDATE to enforce atomic entity updates.