fix(backend): update page atomically in insertBlockIntoPage
Previously, the insertBlockIntoPage GraphQL resolver would get the page
from the database and then update it in separate transactions. This
caused a bug where concurrent requests to this resolver would conflict
and not all new blocks would be present in the updated page. This
commit uses the new getAndUpdateEntity DBAdapter function from the
previous commit to perform the page update atomically.