build(backend): optionally view StatsD metrics with new Docker config
Adds a docker config to backend/statsd containing a barebones StatsD
server which just outputs metrics to the console. By default, this
container is not run when with "yarn serve:hash-backend". To enable
the StatsD server and metric reporting in the API, use
"yarn start:docker-with-statsd" from the backend package.