fix(backend): insert into incoming_links when updating versioned entity
When a new version of an entity is created, we insert the appropriate
references into the incoming_links table. This was causing a bug where
the initial update to a versioned, say Text entity, would be okay, but
subsequent updates would not update parent entities recursively because
the references were not present in the incoming_links table.