fix(be): ensure only latest entity versions are updated in chain
Updating a versioned entity causes a chain of updates to propagate
up the entity's ancestor sub-graph. When performing this update chain,
we need to ensure that only the latest version of the entity is
updated. This commit fixes two bugs related to this issue:

  1. New versions being created on all previous versions of an entity
  2. Concurrent transactions modifying the same entity causing forks
     in entity history.

Bug 1. is a relatively easy fix, and was just an oversight in the
previous code. I fixed this by selecting only the latest entities when
performing the recursive update on parents.

Bug 2 is more subtle. It is easiest to understand with an example.
Let's say we have 3 entities: A, B & C each with a single version
A0, B0 and C0. C0 links to both A0 and B0. There are two concurrent
transactions, one updating A and the other updating B.

  1. tx1: creates a new version of A. It sees that A0 is the latest,
     so creates a new version A1.
  2. tx2: creates a new version of B. It sees that B0 is the latest,
     so creates a new version B1.
  3. tx1: sees C is a parent of A0 with C0 the latest version so
     creates a new version C1 linking to A1.
  4. tx2: also sees that C is a parent of B0, but since tx1 has not
     commited yet, C0 is still the latest version. It creates a new
     version of C, C1'. But, now there is a fork in the version history
     of C.

To fix the forking issue described above, before starting to update
an entity, we acquire a Postgres advisory lock on the entity ID.
If a transaction has an entity locked, other transactions wanting
to update the same entity must wait until the lock is released. Now, 
the sequence of events becomes:

  1.  tx1: acquires lock on entity A.
  2.  tx2: acquires lock on entity B.
  3.  tx1: creates new version A1.
  4.  tx2: creates new version B1.
  5.  tx1: sees C is a parent of A0. Acquires lock on entity C
  6.  tx2: sees C is a parent of B0. Tries to acquire lock on entity C
      but cannot. tx2 waits here until it can acquire the lock
  7.  tx1: sees C0 is the latest version of C. Creates a new version C1.
  8.  tx1: commits & releases locks on A and C.
  9.  tx2: acquires lock on entity C.
  10. tx2: sees C1 is the latest version of C. Creates a new version C2.
  11. tx2: commits & release locks on B and C.

The entity locks ensure that the version history remains linear.

Changes:
  - Modify getParentIds function to return parent entityIds instead
    of entityVersionIds
  - Get the latest version of the entity's parents, and perform the
    recursive update on those.
  - Add acquireEntityLock function which uses pg_advisory_xact_lock
    to acquire a lock on a given entityId.
  - Acquire locks on all entities in the update chain.