feat(be|shared): allow links on entityId, not just entityVersionId
Modifies the __linkedData field to allow for linking on the fixed
entityId, rather than linking to a specific entityVersionId. When an
entity links to another by entityId, the properties resolver will
retrieve the latest version of that entity. Links may still be
made to a specific entity version by specifying the entityVersionId.

With this change, we no longer need to recursively update the ancestor
sub-graph of an entity when performing an update. This significantly
reduces the number of rows written to the database.

This commit also changes how pages and blocks are created in the GraphQL
resolvers to link on entityId rather than entityVersionId.

Changes to packages/hash/shared:
- Use entityId on block and entity in calculateSavePayloads