chore: rename hash/backend package to hash/api & related workspace
Renames the the hash/backend package to hash/api since API is a more
descriptive name of this package now that we have multiple backend
services (e.g. realtime and search-loader). The package is now
called @hashintel/hash-api and the workspace shortcut in the root
package.json is called ws:hash-api