[BE] Search endpoint (internal-pr-352)
* refactor(api): getSystemTypeLatestVersion should never return undefined

System types should always be present in the datastore. If one is not
found then its a non-recoverable error and suggests the system has
not be initialized correctly or has become corrupted. Therefore, it
makes sense to change the return value of the getSystemTypeLatestVersion
to return Promise<EntityType> instead of Promise<EntityType | undefined>

* refactor(search): move SearchAdapter interface & OpenSearch impl to backend-utils

* feat(search): add clear-search command and call it in seed-db

Adds a clear-opensearch.ts script to the search-loader packages which
clears all data in OpenSearch. We call this script in the seed-db
command to refresh the search index alongside the datastore.

* feat(api): add getSystemAccountId function to DBAdapter

* feat(api): add getAncestors and getChildren functions to DBAdapter

* refactor(api): use GracefulShutdown and add http-terminator dependency

Adds a http terminator to the API's HTTP server which correctly closes
all active connections. Also, switches to using the GracefulShutdown
class from backend-utils.

* feat(api): add searchPages API resolver and connect to OpenSearch

- Adds a searchPages resolver to the GraphQL API
- modifies the docker-compose file to add environment variables for
  connecting to OpenSearch.

* fix: run yarn fix:deduplicate

* update with changes requested in PR

* feat(api): throw BadUserInput in searchPages if query is empty

* refactor(be-utils): change OpenSearchConfig field name to maxConnectionAttempts

* change fn name to getAncestorReferences

* throw error if getAncestorReferences depth is less than 1

* uniq on getAncestorReferences return value

* prettier

* fix clear-opensearch script when index does not exist

* fix uniq in getAncestorReferences.ts

* remove redundant type assertion .map

* Make auth optional in OpenSearch adapter