Move `execution::package` to `execution::package::simulation`
When adding experiment packages into this crate as well it's confusing to have them alongside simulation packages. Thus, this commits moves the sim-packages to a submodule and updates the imports to that. If we decide to move out packages to an independent crate, this would only have an `experiment` and `simulation` submodule.