Move experiment packages to the `execution` crate (#536)
* Move `execution::package` to `execution::package::simulation`

When adding experiment packages into this crate as well it's confusing to have them alongside simulation packages. Thus, this commits moves the sim-packages to a submodule and updates the imports to that. If we decide to move out packages to an independent crate, this would only have an `experiment` and `simulation` submodule.

* Move `ExperimentControl` to `execution`

* Move experiment package update comms to `execution`

* Ignore clippy lint on `ExperimentControl`

* Move experiment package control comms to `execution`

* Move `ExperimentPackageComms` to `execution`

* Move `ExperimentName` and `ExperimentPackageConfig` to `execution`

* Make `ExperimentPackage` independent of `ExperimentConfig`

`ExperimentConfig` isn't used for experiment packages. Each experiment package has its own config, so if further configuration is needed, it should be added there.

* Move `ExperimentPackage` to `execution`

* Adjust package paths

* Improve documentation on the package system