Rename functions and add `attach_provider`
This commit applies the changes listed in the PR to the error library. It renames the `wrap_err` and `provide_context` methods to `attach_message` and `change_context`. The documentation were adjusted accordingly.
In addition, it adds `attach_provider`, so the internal representation in `Frame` changes a bit. The impact on existing usages is almost zero (besides the renaming).
The macros were adjusted and documentation were improved.