Fix utoipa component generation
Utoipa requires `#[component(value_type)]` in order to generate the OpenAPI spec. This implies, that we cannot use a generic `PersistedOntologyType<T>` for ontology types. This is not a pretty solution, but we expect, that the implementation for `Read<T: OntologyType>` will change soon as the query builders for ontology types will differ. This applies the changes a little b it earlier than expected, but for now this should be fine.