Change context to shared reference in `Resolve`
`Send` is only implemented for unique references when it is a mutable reference to a shared reference. We require internal mutability to use the context as a cache. Changing this to a shared reference makes the functions on `Read` a lot easier and also avoids the need to implement `PostgreContext` for `&T`