Prepare design system for publishing (#1961)
## 🌟 What is the purpose of this PR?

This PR improves the compatibility of `@local/design-system` with it being published on npm. Because we plan to add a build step, we should avoid imports from `@local/design-system/src/*` because a built version will need it to be `@local/design-system/dist/*`. We could switch over to live incremental rebuilds, it it might be best to tweak `package.json` before publishing for a slightly better DX. More PRs to follow, including https://github.com/hashintel/hash/pull/1953.

I renamed `src/ui.ts` to `src/main.ts` given the opportunity.

<!-- Explain, at a high level, what this does and why. -->
<!-- Use the 'What does this change?' section to list more specific implementation details. -->

## 🔗 Related links

<!-- Add links to any context it is worth capturing (e.g. Issues, Discussions, Discord, Asana) -->
<!-- Mark any links which are not publicly accessible as _(internal)_ -->
<!-- Don't rely on links to explain the PR, especially internal ones: use the sections above -->

- [Asana task](https://app.asana.com/0/1203543021352041/1203443872976464/f) _(internal)_

## ❓ How to test this?

1.  Try importing from `"@local/design-system/src/whatever`
1.  Observe ESLint error saying `Please import from @local/design-system instead`