Attempt to fix linting job in CI (#1977)
## 🌟 What is the purpose of this PR?

In various PR the Rust CI linting job is failing as `cargo make` is supposed to be installed, but it's not. This PR tries to solve this by not caching the tools (which is probably a good idea anyway). If this PR is able to be merged, this issue should be solved, otherwise, it's very likely hitting the same bug.

## 🔗  Related links

This error happened in
- #1800
- #1973 
- #1976