`error-stack`: remove `owo-colors` as dependency (#1980)
## 🌟 What is the purpose of this PR?

Remove `owo-colors` as a dependency. This means we can push the enable `colors` support without an extra feature! `owo-colors` is being replaced with a vendored solution that does the same thing, only more correct.

This also implements the suggestion from https://github.com/hashintel/hash/pull/1800#discussion_r1089753480

### Why remove `owo-colors`?

Don't get me wrong, `owo-colors` is a great crate, but for our purposes is a bit too heavy. 

Initially, we chose `owo-colors` because of the convenience, and we thought that it might provide Windows support (my bad 😅) and automatic detection of color support.

In 0.3, we have removed automatic detection support and discovered that `owo-colors` does not - in fact - provide Windows support. In that case, it is pretty easy to create an alternative with a similar level of convenience, and after going on an adventure 🥾  with @TimDiekmann through the lands of Windows 11 terminals and discovered that pretty much every built-in terminal either supports ANSI escape sequences natively out-of-the-box, or it can be enabled (on older installations).

### Correctness of `owo-colors`

What about correctness? After looking through snapshots I have discovered that for style attributes like `bold` or `italic` `owo-colors` actually does the wrong thing when resetting the styles. I will be posting a follow-up PR for `owo-colors`, but due to it being passively maintained I think there's a low chance it will be in time for 0.3 anyway.

![image](https://user-images.githubusercontent.com/7252775/215569683-2bf9bd12-7076-4fcc-a7c1-f2c2b3545767.png)

The problem: when using style attributes like `bold` (`\e[1m`) or `italic` (`\e[3m`) when you want to end the portion of text there are two options:
1) nuclear: `\e[0m`, this removes all formatting (`owo-colors` uses this)
2) 🎩 gentleman: `\e[22m` (not `bold`), `\e[23m` (not `italic`), instead of removing and resetting everything it just disables bold or italic respectively, allowing for nested styles and even colored text in italic text (not possible with the nuclear option)

We also compress as much as possible, meaning that `bold italic` is not `\e[1m\e[3m` but instead `\e[1;3m`. Small optimization, but a nice one c:

## 👂 Open Questions

I went ahead and just implemented all ANSI escape sequences, they are currently lacking tests (coming soon), but most are currently unused. Should we remove the ones that are not needed or future-proof the solution?

Do we want to expose the coloring methods through `error-stack`?

> IMO, no, if we want to go ahead with something like that, it should go in a separate crate

## 🔗 Related links

## 📹 Demo