Add new `OntologyMetadata` enum that contains variants for Owned and External (#2010)
## 🌟 What is the purpose of this PR?

With #2004 we added support for external metadata to the database. This requires new handling on the Graph API to properly query for that metadata.

For now, only the logic to query the database is implemented, external metadata is not exposed yet to the TS backend. For this, only the `skip_serialize` has to be removed from the added enumeration.

## 🔗 Related links

- [Asana task](https://app.asana.com/0/1203543026679380/1203843814258952/f) _(internal)_

## 🔍 What does this change?

- Split `OntologyElementMetadata` into `OwnedOntologyElementMetadata` and `ExternalOntologyElementMetadata`
- Adjust structural queries to query for the `additional_metadata` column
- Remove old references to the `owned_ontology_ids` table