Define full block names in package.json files (#2024)
## 🌟 What is the purpose of this PR?

This PR enables upcoming updates to the block publishing process. It defines block names in `package.json` → `"blockprotocol"`.

The new values of `"name"` include author and are different to the Yarn workspace names (`@blocks/*`). The only block in which I don’t update `"name"` is _Code_, in which it is equal to `"@hash/code"` already:

https://github.com/hashintel/hash/blob/4a18af145a4e20cda5b7026576256cda8c85d104/blocks/code/package.json#L45-L52

<!-- Explain, at a high level, what this does and why. -->
<!-- Use the 'What does this change?' section to list more specific implementation details. -->

## 🔗 Related links

<!-- Add links to any context it is worth capturing (e.g. Issues, Discussions, Discord, Asana) -->
<!-- Mark any links which are not publicly accessible as _(internal)_ -->
<!-- Don't rely on links to explain the PR, especially internal ones: use the sections above -->
- [Asana task](https://app.asana.com/0/1203358502199087/1203388867587773/f) _(internal)_


## 🧪 How to test this?

- See tests and preview deployment in https://github.com/blockprotocol/blockprotocol/pull/947

There is a chance that HASH app may have some issues because of the new metadata. We will see this after merging https://github.com/blockprotocol/blockprotocol/pull/947. The chance of this is small though because the Code block had `"@hash/code"` as `"name"` already.