Merge `interval-ops` into `graph` crate (#2027)
## 🌟 What is the purpose of this PR?

In order to implement traits on the interval-related types from the graph, those needs to be in the same crate. The other possibility is to move the time-related things to the `interval` ops and add dependencies on `serde`, `utoipa` and `postgres`, which defeats the purpose of having a small interval-library. In the future, we may want to move out the identifiers as a whole, but this is currently not worth doing.

## 🔗 Related links

This is part of [forbidding the "unbounded" bound in temporal axes](https://app.asana.com/0/1202805690238892/1203791431510341/f) and [renaming time related types](https://app.asana.com/0/1203543021352041/1203783495017458/f), but was moved to a separated PR to avoid huge diffs.