Make `Interval` a struct (#2028)
## 🌟 What is the purpose of this PR?

With upcoming tasks we require more interval with different bound types. Implementing `Interval` on all of them is annoying and error-prone. A good example is that we currently don't validate, if `lower <= upper` (a bug no one noticed yet 🙃). By making `Interval` a struct instead, this will enable us to use it everywhere.

## 🔗 Related links

This is part of [forbidding the "unbounded" bound in temporal axes](https://app.asana.com/0/1202805690238892/1203791431510341/f) and [renaming time related types](https://app.asana.com/0/1203543021352041/1203783495017458/f), but was moved to a separated PR to avoid huge diffs.

With the linked tasks I plan to introduce a few aliases for the different intervals to avoid writing the bounds each time, this was captured as TODO.

---
- To see the specific tasks where the Asana app for GitHub is being used, see below:
  - https://app.asana.com/0/0/1203936439791005