Replace Graph initialization-stop-cap-solution to use the type fetcher (#2017)
## 🌟 What is the purpose of this PR?

In order to properly implement the type fetcher, it's required that it's set up properly in external services.
This PR adds the `hash-type-fetcher` binary to the graph Docker image and runs it from the external services. To make sure the deployment is working, it's used to replace the current stop gap solution to insert primitive datatypes but instead it fetches it from blockprotocol.org.

## 🔗 Related links

- [Asana task](https://app.asana.com/0/1203623179643281/1203882684071078/f) _(internal)_

## 🚫 Blocked by

- https://github.com/hashintel/hash/pull/2038

## 🔍 What does this change?

- ~~Add type fetcher binary to Dockerfile~~ https://github.com/hashintel/hash/pull/2038
- ~~Use rust-tls instead of OpenSSL to avoid the dependency on glibc in the Dockerfile (The error messages on alpine on link errors are not helpful at all ("file does not exist"), by luck I've seen the OpenSSL dependency and removed it)~~ https://github.com/hashintel/hash/pull/2038
- ~~Run the type fetcher from external services~~ https://github.com/hashintel/hash/pull/2038
- Implement `FetcherPool`, which is a wrapper over another `Pool`, but also holds information for the type fetcher
- Provide `FetcherStore`, but don't implement `Store` on it yet as this is better suited as a follow-up after some clean-up of the `Store`, which would otherwise result in a huge diff. `FetcherPool<P>` currently acquires `P::Store` instead of `FetcherStore<P::Store>` but added a commented-out section on how this will be done in a follow-up.
- To test this functionality, I replaced the stop-cap-solution we currently use by a function, which uses the type fetcher instead.

## 📜 Does this require a change to the docs?

In order to run the graph we still use `cargo make deployment-up`, which now also covers the type fetcher.

## ⚠️ Known issues

Currently, `Store` requires _all_ different stores to be implemented on that type and `Pool` requires the associated type to implement `Store`. This is a very unfortunate situation and requires some refractoring to sort out. Also I have some issues with `#[async_trait]` as it requires `Send` and cannot capture the inner lifetimes implicitly. When removing `#[async_trait]` or the `Send` bound (`#[async_trait(?Send)]`), `axum` is not able anymore to use methods from those traits. There is a way to work around this, but this will uglify the `Store` trait even more. I will tackle this in a follow-up PR.

## 🐾 Next steps

Dynamically use the type fetcher when creating or updating an ontology type

## 🛡 What tests cover this?

The `type-fetcher` is used to fetch built-in data types from blockprotocol.org. Any test, which uses those built-in types, will also cover this implementation. 

## ❓ How to test this?

Spin up the external services and check if everything works as before