Properly run `clippy` on `hash-graph` (#2073)
## 🌟 What is the purpose of this PR?

We had the `config.toml` in `.config` instead of `.cargo` so clippy did not run recently. This fixes this issue and also makes sure, that clippy will be invoked for integration tests.

## 🔗 Related links

- [Asana task](https://app.asana.com/0/1203193968491715/1203958456055101/f) _(internal)_

## 🔍 What does this change?

- [Address clippy lints](https://github.com/hashintel/hash/commit/c7887f84e267546331af65b2f218aa1b51a97ef4)
- [Run linting on integration tests as well](https://github.com/hashintel/hash/commit/3393d302e4653b2ecfa1ec8ee85dd08cf83f448a)
- [Make integration tests part of the workspace](https://github.com/hashintel/hash/commit/2a48af621b89ea32cc91217c94059eb620bb7142)
- [Apply same structure to benches](https://github.com/hashintel/hash/commit/5d3389ee6646c2da6fd60bf58a72373a13e87aa4)

## 🐾 Next steps

The approach on adding a parameter in `Makefile.toml` worked pretty well. We should use this pattern more often instead of overwriting everything.