(Mostly) Migrate HASH to BP 0.3 (#2072)
## 🌟 What is the purpose of this PR?

For quite a while now we've been iterating upon the Block Protocol spec natively within HASH. We've been testing out a variety of changes to the Type System, the Graph Module, etc. With an incoming launch of a new version of the Block Protocol we want to put the two projects back in sync (mostly).

This PR upgrades the versions of the Block Protocol packages to the newest canary versions which contain [most of the changes being made on the Block Protocol 0.3 branch](https://github.com/blockprotocol/blockprotocol/tree/0.3). In doing so, we have removed a lot of our early implementations of `Subgraph` traversal and now opt for the official ones we ported to the Protocol repo. We also fix a large quantity of mismatches in shapes of data (and namings of fields) that have come about of upstream changes.

We do however, want to maintain more information than the Protocol specifies, including things like provenance metadata, greater information captured inside `EntityId`s, distinct metadata for owned Ontology types, etc. These differences in types caused this PR to be much, much, bigger than originally hoped for, and even getting this in required a lot of prior work to remove some pre-existing cyclical dependencies and package management problems. (See #2040, #2048, #2059).

As such, this PR is absolutely _**massive**_, it was near impossible to separate it into components that were working, you may see that some commits have _(Broken Interim State)_ in their messages, this is to indicate that the app was broken for that commit. Commits without that message should still work as expected. While it is possible to therefore split this PR into a few PRs at those various inflection points, they won't make much semantic sense as chunks of work, and the additional overhead costs of multiple PR reviews, conflicts, etc. mean that a giant PR makes the most sense in this instance.

## 🔗 Related links

- [Asana task](https://app.asana.com/0/1202805690238892/1203701650452068/f) _(internal)_
- [Asana task](https://app.asana.com/0/1203358502199087/1203756383156843/f) _(internal)_
- [Asana task](https://app.asana.com/0/1203543021352041/1203783495017458/f) _(internal)_
- [Asana task](https://app.asana.com/0/1202805690238892/1203791431510341/f) _(internal)_

## 🚫 Blocked by

N/A

## 🔍 What does this change?

Effort has been made to keep individual commits well encapsulated, and clear, and as such the review should be a fairly logical linear process (making bigger ones that involve a tonne of import renames fairly easy to speed through). As such, please look through the commits to understand, and feel free (obviously) to open review comments where it is not clear.

## 📜 Does this require a change to the docs?

- Weirdly enough, I don't think this gigantic PR actually does. We'll probably want to follow-up by renaming `hash-subgraph` to something that better encapsulates what it is, and probably adding rules around its usages (to avoid directly calling `@blockprotocol/graph` in most instances for example)

## ⚠️ Known issues

(All are linked to their relevant threads opened in GitHub)

- [The Subgraph Types are not assignable](https://github.com/hashintel/hash/pull/2072/files#r1106090797)
- [The compatibility of the message callbacks aren't checked](https://github.com/hashintel/hash/pull/2072/files#r1106014332)
- [The message handlers we _do_ implement aren't actually necessarily implemented correctly, or aren't compatible](https://github.com/hashintel/hash/pull/2072/files#r1106017238)
- [There's an issue with the compatibility of the `MergeReturn` type](https://github.com/hashintel/hash/pull/2072/files#r1106020367)
- A portion of the `page-creation` Playwright test was disabled in a previous PR due to the code block being broken, this is still the case

## 🐾 Next steps

- Decide on any actions to take on the issues described above
- Migrate the callback implementations to be compatible
- Fix the blocks that are in the HASH repo, especially the code block so it can be re-enabled within the Playwright tests

## 🛡 What tests cover this?

- `tsc`
- `eslint`
- `test:backend-integration`
- `test:playwright`
- The HASH graph test suite

## ❓ How to test this?

Aside from checking CI passes, please run the app and ensure it all works as usual. It should basically be non-visible change to the end-user.

## 📹 Demo

N/A (would just be a plain video of the app working)