Make fields on common structs `pub` (#2081)
## 🌟 What is the purpose of this PR?

For the same reason as https://github.com/blockprotocol/blockprotocol/pull/987 this PR makes structs, which are only a collection of information (like `OntologyVertexId`, or `EntitMetadata`) publicly accessible. With the linked PR it's also possible to properly implement `From` implementations for `VersionedUri`.

These changes makes a few accesses a little bit more verbose when otherwise chaining is possible, however, in other cases, this greatly improves readability, especially when type hints are available.

I planned to do these changes as a drive-by to my next PR but have split the changes off to make reviewing easier.

## 🚫 Blocked by

- https://github.com/blockprotocol/blockprotocol/pull/987

## 🔍 What does this change?

Please see the list of commits to see the changes

This PR also implements `FromSql` for more types, which previously we didn't because from just looking at the code (without typehints), it was not possible to see, if the parameters returned from `row.get(N)` were passed to the correct location. With these changes it's now required to type the parameter explicitly, so it's easy to check if a call was made correctly (the parameter passed to `get()` can't be checked at compile time). 

Example:

```diff

< let owned_by_id = OwnedById::new(row.get(owned_by_id_index));
< let entity_uuid = EntityUuid::new(row.get(entity_uuid_index));
< let updated_by_id = UpdatedById::new(row.get(updated_by_id_index));
< 
< Ok(Entity::new(
<     properties,
<     link_data,
<     EntityRecordId::new(
<         EntityId::new(owned_by_id, entity_uuid),
<         EntityEditionId::new(row.get(edition_id_index)),
<     ),
<     EntityVersion {
<         decision_time: row.get(decision_time_index),
<         transaction_time: row.get(transaction_time_index),
<     },
<     entity_type_uri,
<     ProvenanceMetadata::new(updated_by_id),
<     // TODO: only the historic table would have an `archived` field.
<     //   Consider what we should do about that.
<     row.get(archived_index),
< ))
---
> Ok(Entity {
>     properties,
>     link_data,
>     metadata: EntityMetadata {
>         record_id: EntityRecordId {
>             entity_id: EntityId {
>                 owned_by_id: row.get(owned_by_id_index),
>                 entity_uuid: row.get(entity_uuid_index),
>             },
>             edition_id: row.get(edition_id_index),
>         },
>         version: EntityVersion {
>             decision_time: row.get(decision_time_index),
>             transaction_time: row.get(transaction_time_index),
>         },
>         entity_type_id,
>         provenance: ProvenanceMetadata {
>             updated_by_id: row.get(updated_by_id_index),
>         },
>         archived: row.get(archived_index),
>     },
> })
```