Make `utoipa` definitions more readable (#2106)
## 🌟 What is the purpose of this PR?

We currently have still a lot of `OneOf` and more unreadable things in our `api.ts` file. This makes it very unreadable. By only tweaking a few small bits on the Graph API we are able to greatly improve readability. One example is removing `OntologyOutwardEdgesOneOfKindEnum` and just using (the already defined) `OntologyEdgeKind` enumeration.

This also fixes a naming issue on the `*Edges` enumeration and removes the trailing `s`.

## 🚫 Blocked by

- #1854