Generalize edge creation in the graph to abstract over `EntityIdWithInterval`s (#2110)
## 🌟 What is the purpose of this PR?

Currently, our representation of edges in the Graph is very limited. We only store a generic `OutwardEdge` struct, which holds exactly one base id and one revision id, merging the intervals for overlapping edge lifespans is only very hard to achieve. This introduces a few small crates to identify the `VertexId` and the `EdgeEndpoint` to abstract over them. A generic `AdjecencyList` was added, which can handle the insertion logic depending on the required edge layout.

This PR does not introduce the union of intervals but provides an entry point for this, where the logic can simply be added in a follow-up PR.

## 🔗 Related links

- [Asana task](https://app.asana.com/0/1203623179643287/1203998818778056/f) _(internal)_

## 🚫 Blocked by

- #1854 
- #2106 
- #2109 

## 🔍 What does this change?

I encapsulated every change in fine-grained commits. Please see the commit list.

## 🐾 Next steps

Merge overlapping edge endpoints ([Asana task](https://app.asana.com/0/1201095311341924/1203399924452451/f))