Allow running external services in "offline" mode (#2126)
## 🌟 What is the purpose of this PR?

When developing locally it's annoying that an internet connection is required to use the type fetcher. The type fetcher has to pull data from `blockprotocol.org` for primitive data types and entity types.

This PR adds the option to run external services without the type-fetcher, which currently falls back to the stopgap solution we still have in the Graph

## 🔗 Related links

- [Slack thread](https://hashintel.slack.com/archives/C02TWBTT3ED/p1677348564924359) _(internal)_

## 🔍 What does this change?

- Move the `type-fetcher` to another compose file
- Use the above compose file as an override configuration for the default behavior and adjust `package.json` to pick that up automatically
- Add a configuration to not run the type fetcher at all and make this accessible with `yarn external-services-offline`

## 📜 Does this require a change to the docs?

A small section to the readme was added

## ⚠️ Known issues

This is only a hotfix. To properly support an offline mode, this requires more programming and design work.

## ❓ How to test this?

Test that `yarn external-services up --wait --build` does work as before
Test that `yarn external-services-offline up --wait --build` does work without starting the `type-fetcher` service