`deer`: `deserialize_optional` (#2071)
## 🌟 What is the purpose of this PR?

I did not want to include this initially, being afraid to be too be too Rust centric and keeping things as generic as possible.

The implementation of the remaining core types has shown that at least a special deserialize method (`deserialize_optional`), similar to [`deserialize_option`](https://docs.rs/serde/latest/serde/trait.Deserializer.html#tymethod.deserialize_option) is needed.

`deserialize_optional` can be considered unique, as it will only ever call three different visitor functions:
* `visit_none`
* `visit_null`
* `visit_some`

`visit_some` takes the `Deserializer` as an argument and enables the possibility of deserializing inner types.

Why is this needed?

The initial implementation of `Option<T>` uses `deserialize_any`, but this isn't sufficient, as only self-describing formats like `JSON`, but not formats like `bincode` can make use of the method, meaning that `Option<T>` (and other types that rely on the fact that a value is either present or not) cannot be used in such circumstances. This is bad. Therefore, a unique deserialize method and visit function have been added.

## Differences between `serde` and `deer`

To ensure that all variants are met and due to the unique nature of `deserialize_optional`, the `deserialize_optional` takes an `OptionalVisitor` instead, which is the same as a `Visitor` except that it only has three visitor methods. This ensures a clear separation of concerns and only allows the invocation of `visit_some` in this particular context (which should always be the case anyway). This means that `deserialize_any` cannot invoke this `Visitor`, but this is by design because `deserialize_any` cannot know if the requested `Visitor` wanted to see if the value exists and most implementations (at least for `serde` AFAIK) do not implement `visit_some`.