feat: parse path instead of symbol
This is necessary to allow for module structures in the future / no collision import, e.g.

`core::ifElse` or `core::let` can always be referred to and while `let` and `ifElse` are used as variables.

```
(“core::let” “let” 12)
(“core::+” “let” “let”)
```

this is valid.

This is because we have no concept of identifiers, just functions.